test: add operator get tests with authorization
